# crates/puddle-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "puddle",
    "description": "puddle's local API: pending requests, rules, audit, settings, consents and the event stream. Listens on 127.0.0.1 only. Every request needs `Authorization: Bearer <token>` (from the connection file), a `Host` of `127.0.0.1:<port>` or `localhost:<port>`, and no foreign `Origin`.",
    "contact": {
      "name": "Tijs van Kampen",
      "email": "puddle@tijsvankampen.be"
    },
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/consents": {
      "get": {
        "tags": [
          "consents"
        ],
        "summary": "Every consent and its state.",
        "operationId": "get_consents",
        "responses": {
          "200": {
            "description": "the consents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consents"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/consents/{kind}": {
      "put": {
        "tags": [
          "consents"
        ],
        "summary": "Records the user's answer to a consent prompt, with the terms version shown. puddle stamps\nthe time.",
        "operationId": "put_consent",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "which consent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConsentKind"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "every consent, after the change",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consents"
                }
              }
            }
          },
          "400": {
            "description": "unknown consent kind",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "invalid terms version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "The global settings and the effective values of a workspace without overrides.",
        "operationId": "get_global",
        "responses": {
          "200": {
            "description": "global settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsView"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Replaces the global settings listed in the body; `null` falls back to puddle's default.\nConsents are changed through `/api/consents`.",
        "operationId": "put_global",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the new global settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettingsView"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "a value out of range, or Microsoft's server chosen without consent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/workspaces/{workspace}": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "One workspace's overrides and effective values. A workspace without stored settings has no\noverrides.",
        "operationId": "get_workspace_settings",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "description": "workspace name",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkspaceName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the workspace's settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "invalid workspace name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Replaces one workspace's overrides; `null` inherits the global value.",
        "operationId": "put_workspace_settings",
        "parameters": [
          {
            "name": "workspace",
            "in": "path",
            "description": "workspace name",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WorkspaceName"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the workspace's new settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "invalid workspace name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing or wrong bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "forbidden origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "written by a newer puddle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "421": {
            "description": "Host is not the API's own address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "a value out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "puddle failed; see its log",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, lower case, no trailing period. Never contains a token."
          }
        }
      },
...
      "Consent": {
        "oneOf": [
          {
            "type": "object",
            "description": "Never asked.",
            "required": [
              "state"
            ],
            "properties": {
              "state": {
                "type": "string",
                "enum": [
                  "not_asked"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Agreed.",
            "required": [
              "at",
              "terms_version",
              "state"
            ],
            "properties": {
              "at": {
                "type": "integer",
                "format": "int64",
                "description": "Epoch ms.",
                "minimum": 0
              },
              "state": {
                "type": "string",
                "enum": [
                  "granted"
                ]
              },
              "terms_version": {
                "type": "string",
                "description": "The terms version shown."
              }
            }
          },
          {
            "type": "object",
            "description": "Declined.",
            "required": [
              "at",
              "terms_version",
              "state"
            ],
            "properties": {
              "at": {
                "type": "integer",
                "format": "int64",
                "description": "Epoch ms.",
                "minimum": 0
              },
              "state": {
                "type": "string",
                "enum": [
                  "declined"
                ]
              },
              "terms_version": {
                "type": "string",
                "description": "The terms version shown."
              }
            }
          }
        ],
        "description": "What the user agreed to or declined."
      },
      "ConsentDecision": {
        "type": "string",
        "description": "The user's answer.",
        "enum": [
          "granted",
          "declined"
        ]
      },
      "ConsentKind": {
        "type": "string",
        "description": "Which consent.",
        "enum": [
          "telemetry",
          "crash_reports",
          "vscode_server"
        ]
      },
      "ConsentRequest": {
        "type": "object",
        "description": "Records the user's answer to a consent prompt. puddle stamps the time.",
        "required": [
          "decision",
          "terms_version"
        ],
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ConsentDecision",
            "description": "Agreed or declined."
          },
          "terms_version": {
            "type": "string",
            "description": "The version of the terms shown (a date or a licence URL; 1 to 256 visible ASCII\ncharacters).",
            "maxLength": 256,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "Consents": {
        "type": "object",
        "description": "Every consent puddle asks for.",
        "required": [
          "telemetry",
          "crash_reports",
          "vscode_server"
        ],
        "properties": {
          "crash_reports": {
            "$ref": "#/components/schemas/Consent",
            "description": "Crash reports."
          },
          "telemetry": {
            "$ref": "#/components/schemas/Consent",
            "description": "Usage telemetry."
          },
          "vscode_server": {
            "$ref": "#/components/schemas/Consent",
            "description": "Microsoft's VS Code server and its licence terms."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "What went wrong, as a stable code a client can switch on.",
        "enum": [
          "misdirected_host",
          "forbidden_origin",
          "unauthorized",
          "bad_request",
          "unsupported_media_type",
          "payload_too_large",
          "invalid",
          "not_found",
          "conflict",
          "newer_settings",
          "unavailable",
          "internal"
        ]
      },
...
      "GlobalSettingsRequest": {
        "type": "object",
        "description": "New global settings. Replaces every value listed here; unknown stored fields are kept.",
        "properties": {
          "ui": {
            "$ref": "#/components/schemas/UiPrefs",
            "description": "Preferences for puddle's window."
          },
          "vscode_server": {
            "$ref": "#/components/schemas/VsCodeServer",
            "description": "VS Code server options."
          },
          "workspace_defaults": {
            "$ref": "#/components/schemas/SettingsLayer",
            "description": "Defaults for every workspace."
          }
        },
        "additionalProperties": false
      },
      "GlobalSettingsView": {
        "type": "object",
        "description": "The global settings, and what a workspace without overrides gets.",
        "required": [
          "workspace_defaults",
          "vscode_server",
          "ui",
          "effective",
          "unknown_fields"
        ],
        "properties": {
          "effective": {
            "$ref": "#/components/schemas/EffectiveSettings",
            "description": "The effective values for a workspace with no overrides."
          },
          "ui": {
            "$ref": "#/components/schemas/UiPrefs",
            "description": "Preferences for puddle's window."
          },
          "unknown_fields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Fields in the stored document this puddle doesn't know (written by a newer one); they\nare kept."
          },
          "vscode_server": {
            "$ref": "#/components/schemas/VsCodeServer",
            "description": "VS Code server options."
          },
          "workspace_defaults": {
            "$ref": "#/components/schemas/SettingsLayer",
            "description": "Defaults for every workspace."
          }
        }
      },
...
      "WorkspaceName": {
        "type": "string",
        "description": "A workspace's name: a DNS label (`a-z`, `0-9`, `-`, no leading or trailing `-`), not `tauri`, `ipc` or `asset`.",
        "example": "my-project",
        "maxLength": 63,
        "minLength": 1,
        "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
      },
...
      "WorkspaceSettingsRequest": {
        "type": "object",
        "description": "New overrides for one workspace. Replaces every value listed; `null` inherits.",
        "properties": {
          "overrides": {
            "$ref": "#/components/schemas/SettingsLayer",
            "description": "The overrides."
          }
        },
        "additionalProperties": false
      },
      "WorkspaceSettingsView": {
        "type": "object",
        "description": "One workspace's overrides and effective values.",
        "required": [
          "workspace",
          "overrides",
          "effective",
          "unknown_fields"
        ],
        "properties": {
          "effective": {
            "$ref": "#/components/schemas/EffectiveSettings",
            "description": "What it gets."
          },
          "overrides": {
            "$ref": "#/components/schemas/SettingsLayer",
            "description": "Its overrides; `null` inherits."
          },
          "unknown_fields": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Unknown fields in its stored document (kept)."
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceName",
            "description": "The workspace."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}